+++ src/pulumi/core-dtap/project_config.py
@@ -1,8 +1,7 @@
 import pulumi
 from os import getcwd, getenv
 from pulumi_azure_native.authorization import get_client_config
-from ingenii_azure_data_platform.config import PlatformConfiguration, \
-    SharedOutput
+from ingenii_azure_data_platform.config import PlatformConfiguration, SharedOutput
 
 # Load the config files.
 platform_config = PlatformConfiguration(
@@ -13,6 +12,7 @@ platform_config = PlatformConfiguration(
     default_config_file_path=getenv(
         "ADP_DEFAULT_CONFIG_FILE_PATH", "../../platform-config/defaults.yml"
     ),
+    metadata_file_path=getenv("ADP_METADATA_FILE_PATH"),
     custom_config_file_path=getenv("ADP_CUSTOM_CONFIGS_FILE_PATH"),
 )
 
@@ -31,5 +31,7 @@ DTAP_ROOT = getcwd()
 
 # Outputs
 platform_outputs = {}
+
+platform_outputs["metadata"] = platform_config.metadata
 
 pulumi.export("root", platform_outputs)