+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - [databricks] - Repositories created in Azure DevOps, and joining to Databricks is a manual step
 - [data_factory] - CI/CD pipelines integrated with data Data Factories
 
+### Improvements
+- [outputs] - Share outputs between Pulumi and Ingenii UI using Azure Tables as storage.
+
 # 0.2.3 (2022-01-17)
 
 ### New Features